+++ Modules/Intent.Modules.AspNetCore.OData.EntityFramework/FactoryExtensions/SwaggerConfigExtension.cs
@@ -33,15 +33,14 @@ namespace Intent.Modules.AspNetCore.OData.EntityFramework.FactoryExtensions
             template.CSharpFile.OnBuild(file =>
             {
                 var @class = file.Classes.First();
-                file.AddUsing("System.Linq");
                 var configureSwaggerOptionsBlock = GetConfigureSwaggerOptionsBlock(@class);
+
                 if (configureSwaggerOptionsBlock is null)
                 {
                     return;
                 }
 
-                configureSwaggerOptionsBlock.AddStatement($@"options.SchemaFilter<{template.GetODataJsonSchemaSwaggerFilterName()}>();");
-                configureSwaggerOptionsBlock.AddStatement("options.ResolveConflictingActions(apiDescriptions => apiDescriptions.First());");
+                configureSwaggerOptionsBlock.AddStatement($@"options.DocumentFilter<{template.GetExcludeODataDocumentFilterName()}>();");
             });
         }