+++ Modules/Intent.Modules.AspNetCore.OData.EntityFramework/Templates/TemplateExtensions.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using Intent.Modelers.Domain.Api;
+using Intent.Modules.AspNetCore.OData.EntityFramework.Templates.ExcludeODataDocumentFilter;
 using Intent.Modules.AspNetCore.OData.EntityFramework.Templates.ODataAggregateController;
-using Intent.Modules.AspNetCore.OData.EntityFramework.Templates.ODataJsonSchemaSwaggerFilter;
 using Intent.Modules.Common.Templates;
 using Intent.RoslynWeaver.Attributes;
 
@@ -13,6 +13,10 @@ namespace Intent.Modules.AspNetCore.OData.EntityFramework.Templates
 {
     public static class TemplateExtensions
     {
+        public static string GetExcludeODataDocumentFilterName(this IIntentTemplate template)
+        {
+            return template.GetTypeName(ExcludeODataDocumentFilterTemplate.TemplateId);
+        }
         public static string GetODataAggregateControllerName<T>(this IIntentTemplate<T> template) where T : ClassModel
         {
             return template.GetTypeName(ODataAggregateControllerTemplate.TemplateId, template.Model);
@@ -21,11 +25,6 @@ namespace Intent.Modules.AspNetCore.OData.EntityFramework.Templates
         public static string GetODataAggregateControllerName(this IIntentTemplate template, ClassModel model)
         {
             return template.GetTypeName(ODataAggregateControllerTemplate.TemplateId, model);
-        }
-
-        public static string GetODataJsonSchemaSwaggerFilterName(this IIntentTemplate template)
-        {
-            return template.GetTypeName(ODataJsonSchemaSwaggerFilterTemplate.TemplateId);
         }
 
     }